TilePyramidCreationProcessForGeoTIFF/requirements.txt
#GDAL==3.6.2
matplotlib==3.9.2
numpy==1.26.4
opencv_python==4.10.0.84
oss2==2.18.5
pandas==2.2.3
Pillow==11.0.0
psycopg2==2.9.9
psycopg2_binary==2.9.9
Py6S==1.9.2
Py6S==1.9.2
pyproj==3.6.1
rasterio==1.3.10
scipy==1.13.1
setuptools==69.5.1
Shapely==2.0.6
